Component: Break projectile components out to own files
