+++ nightmare.vcxproj
@@ -128,6 +128,7 @@
     <ClCompile Include="src\mock_ship_junk.cc" />
     <ClCompile Include="src\physics.cc" />
     <ClCompile Include="src\projectiles.cc" />
+    <ClCompile Include="src\projectile_components.cc" />
     <ClCompile Include="src\settings.cc" />
     <ClCompile Include="src\shader.cc" />
     <ClCompile Include="src\ship_space.cc" />
@@ -155,6 +156,7 @@
     <ClInclude Include="src\physics.h" />
     <ClInclude Include="src\player.h" />
     <ClInclude Include="src\projectiles.h" />
+    <ClInclude Include="src\projectile_components.h" />
     <ClInclude Include="src\scopetimer.h" />
     <ClInclude Include="src\settings.h" />
     <ClInclude Include="src\shader.h" />
